If uverflow issue

+++ arduino/display_button_test/display_button_test.ino
@@ -16,7 +16,7 @@ uint8_t red = 255;
 uint8_t green = 255;
 uint8_t blue = 255;
 
-int lastAction = 0;
+unsigned long lastAction = 0;
 
 enum ScreenState {
   IDLE,
